Add camel_service_queue_task().
Adds a GTask to a queue of waiting tasks with the same source object.
Queued tasks execute one at a time from a separate thread in the order
they were added.

This is primarily intended for use by CamelStore, CamelTransport and
CamelFolder to achieve ordered invocation of synchronous class methods.

This will finally fix Camel's long-standing (mis)use of mutexes as a
dispatch mechanism, which doesn't preserve order and isn't cancellable.