CamelStore: Stop using the internal 'folder_lock'.
Use camel_service_queue_task() instead of locking an internal mutex.
This ensures methods are run in the order they were called, and also
makes cancellation of a waiting method immediate instead of blocking
until it acquires the mutex it's competing for.

This requires flip-flopping some synchronous and asynchronous functions,
so that the synchronous function calls its asynchronous counterpart with
help from CamelAsyncClosure.

Removed functions:

  camel_store_lock()
  camel_store_unlock()