CamelGpgContext: Fix broken debugging statements.
Camel likes to use this debugging idiom:

  #define d(x)

and then wraps debugging statements like so:

  d (printf (...));

One is supposed to enable debugging by changing the macro to:

  #define d(x) x

This idiom makes debugging statements hard to grep for, but moreover
they don't get compiled regularly and so they tend to bit rot.

As was the case in CamelGpgContext, where the debugging statements
didn't build because they were still written for Camel's old home-
grown object system.