IMAPX: Handle BODY[HEADER] in FETCH response.
A broken IMAP server (whatever they use at Masaryk University) was
seen having this conversation:

  C: A001 UID FETCH 1:* (RFC822.SIZE RFC822.HEADER FLAGS)
  S: * 1 FETCH (RFC822.SIZE 4220 BODY[HEADER] {523} ...)
  ...

This confuses Camel into thinking it was a request for a full message
and it in turn fails to update the folder summary and issues a warning
about not being able to locate an appropriate FETCH job.

This works around the issue by treating BODY[HEADER] as equivalent to
RFC822.HEADER in untagged FETCH responses.