CamelIMAPXServer: Ping the server after a period of inactivity.
If an IDLE command is in progress, stop and restart the command.
Otherwise issue a NOOP command.  This just lets the IMAP server
know we're still here to (hopefully) avoid being logged off.

This obviates the need for CamelStore's noop() method, which can
now be removed since it was only needed for IMAP.