Bug 722275 - IMAPX: Parser trips on mailbox containing '+'
Mailbox names containing a '+' character were tripping up the parser due
to it being treated as one of a special set of characters to distinguish
identifiers from IMAP syntax.  But I don't think '+' is necessary in the
set of characters.  For continuations, at least, it should always follow
a newline character, with is also a special character.