IMAPX: Use GSubprocess to implement the shell-command setting.
GSubprocess returns separate input/output streams for stdout/stdin,
and therefore can't be wrapped in a CamelStream.  So this (hidden)
feature is broken until I can fully port IMAPX to GIO streams.

Also, GSubprocess is only available in GLib 2.39, so attempts to
use this feature with an older GLib will fail with a helpful error
message.