imapx_write_flags: Fix a typo which flag to skip
The code was surely trying to avoid writing CAMEL_IMAPX_MESSAGE_RECENT
flag back to the server, but the actual behaviour was that each flag
had been ignored, if the CAMEL_IMAPX_MESSAGE_RECENT was set.
This could have an effect of copied read message being marked as unread
in the destination IMAPx folder.