Add CamelIMAPXInputStream.
Replaces the input side of CamelIMAPXStream.

Currently derived from only GFilterInputStream to keep the code changes
to a minimum until I can verify this works, but eventually would like to
derive from GDataInputStream for reading and buffering whole lines.

Unfortunately this requires switching all of IMAPX over at once, so this
is a rather large commit.