ESource: Add functions for password management.
Replaces the EAuthenticationSession password functions.

Makes password management more accessible.  There's been cases where I
wanted to access the keyring from clients or just outside of the normal
authenication workflow, and had to copy the schema for EDS passwords.

All that's needed from the ESource is its UID, so these functions can be
called on unsubmitted "scratch" sources as well as registered sources.

This will make it possible to accept a password during account creation
and preemptively store it in the keyring to avoid a system modal dialog
when first connecting.

New functions:

   e_source_store_password_sync()
   e_source_store_password()
   e_source_store_password_finish()
   e_source_lookup_password_sync()
   e_source_lookup_password()
   e_source_lookup_password_finish()
   e_source_delete_password_sync()
   e_source_delete_password()
   e_source_delete_password_finish()