Claim missing user name when authenticating
When user's CalDAV/HTTP calendar or WebDAV address book had missing
set login user name, the UI did not show anything useful, the worse
it reasked for a password which were not used at all, because libsoup's
session on the factory side failed in authentication due to runtime
check about non-NULL user name. This makes it return "Authentication
required" error to the UI, instead of asking for the password.