[IMAPx] Manage mailboxes and namespaces in CamelIMAPXStore
The mailboxes and namespaces had been managed in CamelIMAPXServer,
which was perfectly fine before the concurrent connections were added,
because each of the connection could know only on part of the two,
which is not enough. This could let to a failure of folder update
(the case I face were when an empty folder had added old messages,
where one connection new there are some messages, but the connection
serving this update thought there are no messages, thus it didn't try
to download anything).