imapx_connect_sync: Use connection manager to connect to the server
This function opened a new connection to the server, but immediately
dropped it, which, apart of waste of resources, can cause trouble
when other parts asked for the connection to the server, but those
using the connection manager, which is the correct way of dealing
with connections to the server now. The previous behaviour can cause
a race condition on the imapx_store->priv->connecting_server, which
led later to runtime warnings and misbehaviour in general.