Bug #728414 - Deliver ESource data changes on idle
The data D-Bus property change is received in a gdbus dedicated thread,
thus any ESource property changes, caused by reload of the data file,
are also delivered in the same gdbus dedicated thread, which is wrong
when there are still left property bindings which update gtk widgets.
Update the ESource data on idle, to workaround this issue.