imapx_register_job: Try reconnect if the server is shutdown/disconnected
There could sometimes bubble "Not authenticated" error up to the UI,
which didn't look nicely and was not necessary, because in the case
of the server being shutdown or disconnected it's easy to reconnect.
The "Not authenticated" error message is still used, only the error
code is different, it tells the caller to reconnect.