[IMAPx] Try to address "Empty cache file" error
This is rather a wild guess. The UI can be filled with the "Empty
cache file" error, which means that the just-downloaded message
from the server had 0-length size. This can happen when the operation
was cancelled by a user in a "correct time", probably, thus let's
try to check for cancelled operation too - just in case.