camel_imapx_conn_manager_get_connection: Ensure only connected servers are returned
It could happen, still, that the CamelIMAPXServer is disconnected or shutdown
in the list of connections and it is being returned, which is not
correct, thus better never return such dead connections. Once
the server's state is set to shutdown it is automatically removed
from the list, but here plays a role thread interleaving or something,
which does the remove slightly later than expected.