[Maildir] g_mkdir() doesn't fail, if the folder already exists
There could happen that only some of the cur/new/tmp triple were
missing, but in that case the Maildir folder could not be opened,
because the g_mkdir() returned failure and a test for EEXIST was
not in this code part. Other code parts have it, thus let's add
it here as well.