Do not delete passwords during authentication session
The stored password was deleted before actual password prompt
was shown, which is not correct, because the reason to fail with
the stored password can be either that the password on the server
changed or some issue with libsecret, not providing the correct
stored password. As it happens sometimes that the libsecret fails,
then it is quite bad to reenter all the passwords only due to this
reason, thus better do not delete the password at all. The worse
thing on the fact that the password is deleted before the actual
password prompt is shown is that the password prompts pile in a queue,
thus even if the source registry process is killed quickly enough,
at the first false password prompt, all the other stored passwords
are already gone. Not talking that the piled password prompts
can be cancelled meanwhile.