Share subprocesses for built-in backends by default
It turned out that opening subprocesses for each single instance
is a very CPU and resource expensive operation, which can be seen
for example when opening a meeting invitation email, which does
a flood of open requests to open each configured calendar (and
a Birthdays & Anniversaries calendar opens some books). Using one
subprocess for a backend type makes it use less resources.