Bug 735523 - EBookSqlite: Support boolean summary field for EContactCert
We don't actually care about summarising the *data*; merely indicating
the *presence* of the field is sufficient. With the new evolution-pkcs11
module, the 'exists(x509Cert)' query is going to get very hot...

This halves the time it takes to query the Intel global addressbook,
with almost quarter of a million entries of whom about 5% have
certificates listed.