Use 90 seconds timeout for Camel connections
The Camel connections used to use timeouts, but they were unset
for IMAPx's IDLE. The outcome was that all other connections, not
only the IDLE, could wait for ever. A better way to deal with it
is to unset timeout on the connection only when necessary, aka only
during the IDLE being issued on the connection, and let all the other
connections timeout when they do not get a response within the limit.