Disabling/enabling Collection master source doesn't update children properly
Calendar, contacts and mail related children sources depend on two
parent's properties, one is for the sub-part enabled state and another
is the 'enabled' property of the parent ESource. Such children
sources listen for change of their sub-part only, which might reflect
incorrect value of the child's enabled state. This could be fixed
on the server side on load of the children, but the change was not
noticed, thus the ESource was not saved and the client side loaded
the ESource with an old value from the disk.