[IMAPx] Circle between tagprefix-es for connections
Even the connection manager didn't use tag prefixes which were used
by other connections, it was using always the same connection tag
prefix for the last connection which was closed, because it was
picking the tag prefix always from 'A'. While it might not be an issue
to close the connection with one tag prefix and start a new connection
with the same tag prefix to the same server, then it looks better
to change also the tag prefix (which can be eventually observed
in a debug log as well).