Should not ask for password of GOA IMAP account
Evolution was asking for a GOA IMAP account password, even it should
take it from GOA, not from a user. The reason was that the function
e_source_credentials_provider_ref_credentials_source() works for
an Authentication extension, which the GOA (and UOA) doesn't use,
thus the correct credentials source was not found. Providing a similar
lookup function which lookups for GOA (or UOA) extension in the parent
sources makes this work properly.