edataserverui: Remove redundant check for an error being set
There is no way this loop can have local_error set before checking
whether it is NULL. Even if the declaration of local_error was hoisted
to the parent scope, all control paths which lead to the loop being
entered would have local_error set to NULL. This might not be the
correct fix, but there is certainly a problem somewhere in this code.
Spotted by Coverity.

CID: #1296548

https://bugzilla.gnome.org/show_bug.cgi?id=749190