[IMAPx] Message moved to real Trash/Junk multiple times
When a real Trash/Junk is set for the account, then the deleted/junk
messages are moved to the respective folder according to the internal
folder's hash table of the message UIDs to be moved. The hash table is
populated on various places, but emptied only on one. There happened
that the move to the Trash/Junk folder emptied the hash table, but
then the moved UID was re-added back during the move operation, thus
the next folder visit and the change save there was re-tried to move
already moved, thus nonexistent, message into the real Trash/Junk
folder, which could fail with error.

This change makes sure that the hash tables of UIDs to be moved to
the Trash/Junk folders contain only existing UIDs.