Provide default implementation of EBackend::authenticate_sync()
This only returns 'success' for any requests to not force each descendant
define the same function. Some of the descendants, like the GTasks calendar
backend, doesn't need this method, because it requires OAuth2 token,
which is received through GOA. Without this a false "failed to open"
error was shown in the UI, with an "Unknown error" description, because
it failed on a runtime constraint.