Do not add Google Tasks source for non-GOA master sources
Google mail account configured directly in the Evolution can have set
also Contacts and Calendar parts, which makes it a collection.
The Google source-registry module adds for each such collection
the Tasks source, which is not correct, because the GTasks require
OAuth2, which is provided only through GOA. Thus do not add the GTasks
source for non-GOA collection sources.