e-source-registry-server: Watch for changes to .source files and reload
It is possible for evolution-source-registry to detect a newly-added
source file before the author of the file has written to it; the
registry will attempt to load the file, find it is invalid, and then
ignore it.

The registry should also listen for changes to files, and reload them
accordingly. This means that the source file will get reloaded after the
author has finished writing to it, and hence the source will get
created.

https://bugzilla.gnome.org/show_bug.cgi?id=751119