[IMAPx] Name CAMEL_IMAPX_ERROR codes
There were used only direct numbers, instead of enum codes. It's easier
to distinguish between malformed server response and other errors with
the codes. For example, when the server returned malformed response for
a GETQUOTAROOT, the IMAPx was reconnecting to the server again and again,
instead of stopping and returning the error.