[IMAPx] Junk/NotJunk message flags not stored server-side
The Junk/NotJunk flags are not standard flags, they are user flags,
thus, when parsing mailbox's permanent flags response, the Junk/NotJunk
flags should be added to the list of supported flags by the server.
This is true for folders which did not contain any Junk/NotJunk marked
messages yet.