[Weather calendar] Location/Unit changes not propagate until re-open
With this, either the Location or the temperature Unit change will
cause refresh of the calendar and will show the values as being
requested without require of the backend close and open. This is doing
full refresh, aka talks to the server, thus it works only when being
online. When being offline, or the weather server being unreachable,
shows stale data anyway, thus it might be fine.