Bug 755735 - Add a systemd user service for each D-Bus session service
This means each service will run in its own cgroup under a systemd
user session, instead of being a child of dbus-daemon that is
considered to be part of dbus.service. It does not cause the
services to be started on login: they are still started on-demand
by bus activation.

The files installation can be disabled with configure
a option --without-systemduserunitdir.