Make sure EDBusServer modules are loaded only once
After recent commit which checks for modules directory changes
the changed modules could be loaded multiple times, which results
in a GType runtime warnings and eventually a crash. This commit
ensures that each module is loaded only once, for a price of not
have loaded the most recent version of the module, which is not
a problem, because it's not the aim of the modules directory
change watcher.