[EDBusServer] Reload the server only if a new module was added
There is no need to reload the server when any of the already loaded
modules was changed or removed, because only newly added modules can
be loaded on the fly, thus avoid the useless reload. That means that
software update still requires manual background processes restart,
while any 3rd-party module installation does not.