[CalDAV] Add Bearer auth to the SoupSession only when using OAuth2
It can be that the server advertises both Basic and Bearer authentication
methods, while the calendar is setup to use the Basic. When the SoupSession
checks for authentications it can pick the Bearer instead, even when
the calendar doesn't have any credentials (OAuth2 token) for it.
Not passing the Bearer authenticator to the SoupSession helps it
to pick the right authenticator.