[IMAPx] Do not change Sent folder for GMail in initial setup
GMail stores sent messages on the server side automatically, thus
saving a sent message there would make only doubled copies of that
message on the server. This doesn't change the Sent folder at all,
thus if users have it already set to the server, or anywhere else,
then the setting is preserved.