[IMAPx] Release idle_lock before disconnecting the cancellable
There could happen a deadlock when one thread was stopping the IDLE thread
while another thread was cancelling the same cancellable in the disconnect
phase. It could be that the stop_idle thread was holding the idle_lock, while
the other thread was holding the GCancellable's global lock (for example
as part of the g_cancellable_cancel() call) and also waiting for the idle_lock
due to the idle_cancellable being cancelled.

Let's unlock the idle_lock before disconnecting the cancellable, thus these
two locks do not overlap.