[CalDAV] Backend not initialized when asked to authenticate
There could happen a race condition where one thread was opening
the calendar, but another thread was processing an authentication
request and that later thread won the race for the busy_lock, but
the backend wasn't initialized, thus the priv->uri was still NULL
and a runtime warning:
> soup_message_new: assertion 'uri_string != NULL' failed
had been printed on the factory console. The reason was that
the collection backend called authenticate for its children. Count
with this case and initialize the backend when needed, rather than
fake-fail or fake-reject the credentials in the authenticate callback,
because the fake response can have side effects on the credentials.