[CamelGpgContext] Provide signer photos when available
This allows show of signers photo, when the public key has any and
the gpg/2 provides it. This behavior is enabled by default, but
it can be disabled by:
   $ gsettings set org.gnome.evolution-data-server camel-cipher-load-photos false

The size of public struct _CamelCipherCertInfo changed, thus a soname version
bump should be done, but as it had been changed recently and no release of
the evolution-data-server had been made meanwhile, then it was not changed.