Introduce an ENetworkMonitor
It's a GNetworkMonitor implementation, a proxy for a real instance,
which allows to change in runtime which implementation of the GNetworkMonitor
will be used, together with a special option to report an always-online
state. This instance can be used by any part, thus the changes in it
influence not only Mail, but also Contacts, Calendars, Memo Lists and
Task Lists. It wasn't easily possible with the raw GNetworkMonitor.

The added GSettings option is:
   org.gnome.evolution-data-server network-monitor-gio-name
and the ENetworkMonitor uses it transparently.