ENetworkMonitor: Implement also the async part of the GNetworkMonitor interface
It seemed like the async part is provided for free and the interface
lets it override for descendants only when needed, but it turned out
that it's not the true, thus the async part of the interface is defined
too.

This exhibited for example when entering the Calendar view, which called
this async part to figure out whether the calendars are reachable, but
as the default implementation only calls the synchronous part, then it
froze the UI for some time, instead of doing things really asynchronously.