Bug 770452 - Stop using camel_pstring_peek()
If a call to this function results in a string being added to the pool, the
caller effectively owns this initial reference, but it has no way to
distinguish between this and the case where a pointer to an existing string
was returned with no reference added. Thus camel_pstring_free() can never be
called on the result, and any strings that are added this way must stick
around in memory forever.

This is unreasonable for a global string pool that is used to hold a large
number of strings, so avoid using this function.