Bug 767821 - Recognize X-GM-EXT-1 as SPECIAL-USE for Gmail IMAP ][
Do not inherit folder types from store's summary, because those are advertised
by the server, while it might not always match what the user has set. In that
case there could be for example two Trash folders highlighted in the UI, while
only one of them would be the correct one.