[Camel] Introduce camel_operation_new_proxy()
It creates a new CamelOperation, which listens for 'cancelled' signal
of the original operation, but not the other way around. That allows
to cancel the operation without influencing the passed-in cancellable.
Other CamelOperation changes (push/pop message and progress) are also
passed to the original cancellable, being it a CamelOperation as well.

That's because there could happen that for example CamelService's
connect and disconnect do cancel the passed-in cancellable on its own,
but that's wrong, because it can cancel any larger operation to which
it belongs, like when going online or offline from UI, on user's request.