Ignore errors when getting message for a body-contains search
The reason is that the other places which require complete message also
ignore any errors and the error was not respected anyway, thus it could
result in a spread of errors on the console. This way one can search
when being offline and the result will be for the available messages.