Mis-filters emails with Mailing List rule in certain situations
When the filter rule value starts with an '@', but the List-Id header
has no '@' part, then the filter changes the value to pre-domain part,
which is an empty string here, and compares it to the List-Id value,
which basically always succeeds. The change here makes sure that there
are never compared empty strings.

Other part of this change enhances filters logging.

This had been reported downstream:
https://bugzilla.redhat.com/show_bug.cgi?id=1450560