Prevent busy-loop opening Google calendar configured in GOA
This is semi-new in master.

The change in e-backend.c prevents constant ping-pong with source registry
when trying multiple times in a row with empty credentials and the backend
returning 'REQUIRED', which can misbehave with OAuth2 tokens.

The change in e-soup-session.c fixes a new issue in master, which
caused such busy loops due to requiring non-empty credentials for
OAuth2 authentication, while it had been needed only for Google
authentication method.