Introduce ESourceRegistryWatcher
It can be tricky to watch for changes in ESourceRegistry, especially
when the user wants to know only about a specific subset of ESources-s.
This new object makes things significantly easier and reliable, without
a need for code duplication. It is currently used in the 'Birthdays & Anniversaries'
calendar, but it can be reused elsewhere too (like in the evolution-alarm-notify).