Bug 772572 - Allow creating Receive-only mail accounts
Added a new e_util_identity_can_send() helper function, to check whether
the identity source has defined a transport source and this transport
source is not 'none' backend. It doesn't check whether the identity
or the transport source are enabled, it tests only the transport source
existence and its backend name.