Warn and set default when sendmail not found during the build
When sendmail is not available when building, the corresponding SENDMAIL_PATH
variable contains useless value. This change makes it possible to define
SENDMAIL_PATH as an environment variable, which then points to
the sendmail binary (like "/usr/sbin/sendmail"), and if not set, then
it is searched for and when even that fails, then a default is used.

There had been also SYSTEM_MAIL_DIR variable, but it was not used
anywhere, thus it's removed now.

The issue with missing sendmail had been reported at:
https://bugzilla.redhat.com/show_bug.cgi?id=1468890