[IMAPx] Not all local changes saved to the server
IMAPx uses camel_message_info_get_folder_flagged_stamp() also to recognize
which messages are flagged (locally changed) and which are which are only
dirty (requires save to summary). The thing is that the initial
folder-flagged-stamp can be zero, which means g_hash_table_lookup()
use here behaving like there is no value, even the UID is there.