[Camel] Speed up message filtering on headers
When a filter had been set on some headers it always caused message
download, even when the CamelMessageInfo contained the headers. This
checks whether headers are available in the CamelMessageInfo and if
so, then it is using them, rather than downloading the whole message.
To have this for every provider, the CamelFolderSummary's
message_info_new_from_headers() automatically stores passed-in
headers into the created CamelMessageInfo.