Internal Google OAuth2 token not refreshed on calendar open
While the token can be refreshed on demand, it was not done due to the code
considering expired tokens as unusable. That's partly true, except the code
can refresh information when the secret is available. The change sets
at least the stored secret even when the token is expired, which makes
it refresh the token on demand, as is expected.