[CalDAV/CardDAV] Improve logic for object load from the server
One change related specifically for Google servers, to not try twice when
the first load fails with "404 Not Found".

The other change is to verify that the collection on the server did change
meanwhile. If it did, then it's possible it was due to the component which
is not in the local cache yes, thus try to load it. If the collection did
not change, then it means the local cache is up-to-date with the server
content, thus the to-be-loaded object is not on the server, thus no need
to poke it and receive failure anyway. This works only if the server
supports getctag extension.