[CalDAV/CardDAV] Append ending slash to URL when missing
The URL set in the ESource is supposed to point to a collection
and as such should end with a forward slash. While some servers
do not care that much, some others can report an error when
the ending slash is missing. This ensures it's always there.

Reported downstream at:
https://bugzilla.redhat.com/show_bug.cgi?id=1505337