[CamelOfflineStore] Incorrect disconnect when going online
When a CamelOfflineStore had been going online, the other calls inside
the function could connect the store, but the function used the connection
state from the beginning of the function (stored in a variable), which
means it could be using obsolete information. Another thing had been that
the disconnect doesn't make sense when the store is going online and
when it is either disconnecting or disconnected already.