Bug 790267 - Expired Google OAuth2 token not refreshed ][
Book/calendar part didn't work properly. Its workflow was:
a) try with empty credentials => fail with CREDENTIALS_REQUIRED response
b) source registry noticed it and asked for credentials, but
   the Google provider returned no credentials when reading them
   from the store (due to expired access token)
c) left the source wait for credentials until they are provided
   by some GUI application
d) in case of Evolution it could show either "Unknown error" or
   "Credentials required" error in the UI, instead of refreshing
   the token.

This changes the b) in a way that when the lookup fails, but it still
can read credentials, only the token is expired, then it'll refresh it
and save, which will connect the book/calendar without a need for a GUI
application with running credentials prompter.