Prevent passing NULL ldap handle into LDAP functions
It can happen that the LDAP backend is disconnected from the server
while it's processing some operation, in which case the check for non-NULL
ldap handle could come after this had been passed to an LDAP function,
which can eventually have an assertion on non-NULL ldap handle to be
passed in. This change handles the NULL ldap handle case gracefully.

This had been reported downstream at:
https://bugzilla.redhat.com/show_bug.cgi?id=1530569